Cleanup – rename files to fit content
